refactor(tf2-ray-test): spilt from horovod
